+++ src/babble/babble.go
@@ -32,14 +32,27 @@ func NewInmemBabble(config *babble_config.Config, logger *logrus.Entry) *InmemBa
 IMPLEMENT CONSENSUS INTERFACE
 *******************************************************************************/
 
-// Init instantiates a Babble inmemory node
+// Init instantiates a Babble inmemory node.
+//
+// XXX - Normally, the Babble object takes a reference to the InmemProxy via its
+// config. Here, we need the InmemProxy to have a reference to the Babble object
+// as well; a sort of circular reference, which is quite ugly. This is necessary
+// because the InmemProxy calls the babble object directly to retrieve the list
+// of validators. We will change this when Blocks are modified to contain the
+// validator-set. cf. work on babble merkleize branch.
 func (ib *InmemBabble) Init(state *state.State, service *service.Service) error {
 	ib.ethState = state
 	ib.ethService = service
 
-	ib.config.Proxy = NewInmemProxy(state, service, service.GetSubmitCh(), ib.logger)
-
 	babble := babble.NewBabble(ib.config)
+
+	inmemProxy := NewInmemProxy(state,
+		service,
+		babble,
+		service.GetSubmitCh(),
+		ib.logger)
+
+	ib.config.Proxy = inmemProxy
 
 	err := babble.Init()
 	if err != nil {